% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{cgAttachments}
\alias{cgAttachments}
\title{Get multiple primary attachments from a class}
\usage{
cgAttachments(class, outDir = "",
  filter = "([PrimaryAttachment]\%20<>\%20\\"\\")", zip = FALSE, un,
  pw, org, base_url = "https://cgweb06.cartegraphoms.com/")
}
\arguments{
\item{class}{class requesting attachments from}

\item{outDir}{save directory defaults to class name, optional}

\item{filter}{filter to be applied to class}

\item{zip}{option to zip the folder}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}

\item{base_url}{API Base URL (defaulted to "https://cgweb06.cartegraphoms.com/")}
}
\value{
folder of saved jpeg images
}
\description{
Get multiple primary attachments from a class
}
\examples{
\dontrun{
cgAttachments("cgFacilities",
    filter = "([Inactive]\%20=\%20false)",
    un = "fakeUn",
    pw = "fakePwd",
    org = "AnytownUSA")
}
}
