% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{cgDelete}
\alias{cgDelete}
\title{Title}
\usage{
cgDelete(class, oid, un, pw, org)
}
\arguments{
\item{class}{cartegraph class or class and attachment ie 'cgSigns_cgAttachmentsClass'}

\item{oid}{Oid of the object you wish to delete}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}
}
\value{
http status
}
\description{
Title
}
\examples{
\dontrun{
cgDelete("cgSignsClass", "123", "fakePW", "AnyTownUSA")
}
}
