% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{cgDf}
\alias{cgDf}
\title{Return dataframe object rom CartegraphA PI}
\usage{
cgDf(class, fields = "", filter = "", un, pw, org)
}
\arguments{
\item{class}{cartegraph class name}

\item{fields}{fields you want returned}

\item{filter}{optional filter}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}
}
\value{
data.frame
}
\description{
Return dataframe object rom CartegraphA PI
}
\examples{
\dontrun{
pools <- cgDf("PoolsClass", "Oid,IDField,NeighborhoodField,PoolCapacityField,PoolTypeField,RetiredField,WaterSourceField", un = "fakeUn", pw = "fakePwd", org = "AnytownUSA")
}
}
