% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{cgPoints}
\alias{cgPoints}
\title{Return Points}
\usage{
cgPoints(class, fields = "", filter = "", un, pw, org)
}
\arguments{
\item{class}{cartegraph class name}

\item{fields}{fields you want returned}

\item{filter}{optional filter}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}
}
\value{
SpatialPointsDataFrame
}
\description{
Return Points
}
\examples{
\dontrun{
signs <- cgPoints("cgSignsClass",
c("MUTCDCodeField", "LocatorAddressNumberField",
"PavementField", "LocatorStreetField", "LocatorCityField",
"MountingFixtureField", "InstalledField"),
'(\%5BMUTCDCode\%5D\%20in\%20(\%22R1-1\%22,\%20\%22W3-D1\%22))',
"fakeUn""fakePwd")
}
}
