% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{cgPoly}
\alias{cgPoly}
\title{Return Polygons}
\usage{
cgPoly(class, fields = "", filter = "", un, pw, org)
}
\arguments{
\item{class}{cartegraph class name}

\item{fields}{fields you want returned}

\item{filter}{optional filter}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}
}
\value{
SpatialPointsDataFrame
}
\description{
Return Polygons
}
\examples{
\dontrun{
cgPoly("cgFacilities",
    filter = "([Inactive]\%20=\%20false)",
    un = "fakeUn",
    pw = "fakePwd",
    org = "AnytownUSA")
}
}
