% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{cgPost}
\alias{cgPost}
\title{Create new record(s) in Cartegraph}
\usage{
cgPost(class, body, un, pw, org,
  base_url = "https://cgweb06.cartegraphoms.com/")
}
\arguments{
\item{class}{cartegraph class or class and attachment ie 'cgSigns_cgAttachmentsClass'}

\item{body}{DataFrame/tibble to be turned into JSON object}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}

\item{base_url}{API Base URL (defaulted to "https://cgweb06.cartegraphoms.com/")}
}
\value{
http status
}
\description{
Create new record(s) in Cartegraph
}
\examples{
\dontrun{
df <- data.frame(IDField = "Sign-50",
     AddressNumberField = 1765)

cgPOST("cgSignsClass", df,"fakeUN", "fakePW", "AnyTownUSA")
}
}
