% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{cgPut}
\alias{cgPut}
\title{Put Data in Cartegraph}
\usage{
cgPut(class, body, un, pw, org)
}
\arguments{
\item{class}{cartegraph class or class and attachment ie 'cgSigns_cgAttachmentsClass'}

\item{body}{DataFrame/tibble to be turned into JSON object}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}
}
\value{
http status
}
\description{
Put Data in Cartegraph
}
\examples{
\dontrun{
df$cgSignsClass <- data.frame(Oid = 242283165,
     AddressNumberField = 4771,
     StreetField = '123 Main Street')

cgPut("cgSignsClass", df,"fakeUN", "fakePW", "AnyTownUSA")
}
}
