% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{cgShapeProcess}
\alias{cgShapeProcess}
\title{**UNTESTED** Process a Spatial Shape to pass to Cartegraph API through POST or PUT. (Tested for Polygons, not tested for Lines or Points)}
\usage{
cgShapeProcess(shape)
}
\arguments{
\item{shape}{Spatial DataFrame to be turned into nested CgShape column}
}
\value{
tibble with cgShape Column
}
\description{
**UNTESTED** Process a Spatial Shape to pass to Cartegraph API through POST or PUT. (Tested for Polygons, not tested for Lines or Points)
}
