% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{cgDelete}
\alias{cgDelete}
\title{Delete a record in Cartegraph}
\usage{
cgDelete(class, oid, un, pw, org,
  base_url = "https://cgweb06.cartegraphoms.com/")
}
\arguments{
\item{class}{cartegraph class or class and attachment ie 'cgSigns_cgAttachmentsClass'}

\item{oid}{Oid of the object you wish to delete}

\item{un}{api username}

\item{pw}{api password}

\item{org}{orginization API ID ie 'PittsburghPA'}

\item{base_url}{API Base URL (defaulted to "https://cgweb06.cartegraphoms.com/")}
}
\value{
http status
}
\description{
Delete a record in Cartegraph
}
\examples{
\dontrun{
cgDelete("cgSignsClass", "123", "fakePW", "AnyTownUSA")
}
}
