% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaData.R
\name{cgMeta}
\alias{cgMeta}
\title{Meta Data}
\usage{
cgMeta(class = "", un, pw, org,
  base_url = "https://cgweb06.cartegraphoms.com/")
}
\arguments{
\item{class}{string (optional) specify class or classes to get metadata for}

\item{un}{string cartegraph username}

\item{pw}{string cartegraph password}

\item{org}{orginization API ID ie 'PittsburghPA'}

\item{base_url}{API Base URL (defaulted to "https://cgweb06.cartegraphoms.com/")}
}
\value{
list
}
\description{
Meta Data
}
\examples{
\dontrun{
 # Entire System Meta Data
 cgMeta()
 # Meta Data for two classes
 cgMeta(c("cgFacilitiesClass", "cgPavementClass"))
 }
}
